input stream extension
renamed iosfrstor to basic_iosfrstor and made it a template class taking a char type and a char traits type.
internally-stored pointer is now a std::basic_ios, which allows it to additionally work on input streams with no loss of performance.

typedef iosfrstor aliases basic_iosfrstor<char> for common use cases.