added more accessors
added get/set accessors for all the stored fields.
this way, if you want the rstor action to e.g. clear the width field, you can just set it to 0 in the iosfrstor obj immediately after creation.

added a get accessor for the stored stream (might be handy to know what it's liked to).